v4/requirements.txt
requests

sentence-transformers 
faiss-cpu
pandas 
openpyxl

fastapi
uvicorn
pydantic
numpy
openai
dotenv
qdrant-client
python-dotenv